+++ database/migrations/2019_04_04_021330_create_preguntas_table.php
@@ -17,7 +17,7 @@ class CreatePreguntasTable extends Migration
             $table->bigIncrements('id'); /*Esta debe ser llave Primaria */
             $table->string('nombre_preg');
             
-            $table->tinyint('numero_con')->unique(); /*Esta debe ser llave foranea */
+            $table->integer('numero_con')->unique(); /*Esta debe ser llave foranea */
 
             $table->timestamps();
         });
